requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
gunicorn==20.1.0
certifi==2022.12.7
chardet==5.1.0
crispy-bootstrap5==0.7
decorator==5.1.1
Django==4.2
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.1
django-js-asset==2.0.0
executing==1.2.0
Faker==18.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
psycopg2
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.0
python-dateutil==2.8.2
pytz==2023.3
rfc3986==2.0.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
django-dotenv
djangorestframework==3.14.0
python-dotenv==1.0.0